src/chiron/core.py
"""Core functionality for Chiron."""

import logging
import structlog
from typing import Any, Dict, Optional
from chiron.exceptions import ChironError, ChironConfigurationError


class ChironCore:
    """Core Chiron functionality with observability and security features."""

    def __init__(
        self,
        config: Optional[Dict[str, Any]] = None,
        enable_telemetry: bool = True,
        security_mode: bool = True,
    ) -> None:
        """Initialize ChironCore.

        Args:
            config: Configuration dictionary
            enable_telemetry: Whether to enable OpenTelemetry instrumentation
            security_mode: Whether to enable security features
        """
        self.config = config or {}
        self.enable_telemetry = enable_telemetry
        self.security_mode = security_mode
        
        # Configure structured logging
        structlog.configure(
            processors=[
                structlog.stdlib.filter_by_level,
                structlog.stdlib.add_logger_name,
                structlog.stdlib.add_log_level,
                structlog.stdlib.PositionalArgumentsFormatter(),
                structlog.processors.TimeStamper(fmt="iso"),
                structlog.processors.StackInfoRenderer(),
                structlog.processors.format_exc_info,
                structlog.processors.UnicodeDecoder(),
                structlog.processors.JSONRenderer(),
            ],
            context_class=dict,
            logger_factory=structlog.stdlib.LoggerFactory(),
            wrapper_class=structlog.stdlib.BoundLogger,
            cache_logger_on_first_use=True,
        )
        
        self.logger = structlog.get_logger(__name__)
        
        if self.enable_telemetry:
            self._setup_telemetry()
        
        if self.security_mode:
            self._setup_security()

    def _setup_telemetry(self) -> None:
        """Set up OpenTelemetry instrumentation."""
        try:
            from opentelemetry import trace
            from opentelemetry.sdk.trace import TracerProvider
            from opentelemetry.sdk.trace.export import BatchSpanProcessor
            from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
            
            # Set up tracing
            trace.set_tracer_provider(TracerProvider())
            tracer = trace.get_tracer(__name__)
            
            # Configure OTLP exporter
            otlp_exporter = OTLPSpanExporter(
                endpoint=self.config.get("otlp_endpoint", "http://localhost:4317")
            )
            span_processor = BatchSpanProcessor(otlp_exporter)
            trace.get_tracer_provider().add_span_processor(span_processor)
            
            self.tracer = tracer
            self.logger.info("OpenTelemetry instrumentation enabled")
            
        except ImportError as e:
            self.logger.warning(
                "OpenTelemetry not available, telemetry disabled",
                error=str(e)
            )
            self.tracer = None

    def _setup_security(self) -> None:
        """Set up security features."""
        self.logger.info("Security mode enabled")
        # Security setup would go here
        # - Input validation
        # - Authentication/authorization
        # - Audit logging
        # - Rate limiting

    def validate_config(self) -> bool:
        """Validate the configuration.

        Returns:
            True if configuration is valid

        Raises:
            ChironConfigurationError: If configuration is invalid
        """
        required_fields = ["service_name"]
        
        for field in required_fields:
            if field not in self.config:
                raise ChironConfigurationError(
                    f"Required configuration field '{field}' is missing"
                )
        
        return True

    def health_check(self) -> Dict[str, Any]:
        """Perform a health check.

        Returns:
            Health check status and details
        """
        status = {
            "status": "healthy",
            "version": "0.1.0",
            "telemetry_enabled": self.enable_telemetry,
            "security_mode": self.security_mode,
            "timestamp": structlog.processors.TimeStamper(fmt="iso")._stamper(),
        }
        
        self.logger.info("Health check performed", **status)
        return status

    def process_data(self, data: Any) -> Any:
        """Process data with observability and security checks.

        Args:
            data: Input data to process

        Returns:
            Processed data

        Raises:
            ChironError: If processing fails
        """
        if self.tracer:
            with self.tracer.start_as_current_span("process_data") as span:
                span.set_attribute("data.type", type(data).__name__)
                return self._process_data_internal(data)
        else:
            return self._process_data_internal(data)

    def _process_data_internal(self, data: Any) -> Any:
        """Internal data processing logic."""
        try:
            # Security validation
            if self.security_mode:
                self._validate_input(data)
            
            # Process the data
            self.logger.info("Processing data", data_type=type(data).__name__)
            
            # For now, just return the data as-is
            # In a real implementation, this would contain the core business logic
            return {"processed": True, "original": data}
            
        except Exception as e:
            self.logger.error("Data processing failed", error=str(e))
            raise ChironError(f"Failed to process data: {e}") from e

    def _validate_input(self, data: Any) -> None:
        """Validate input data for security."""
        if data is None:
            raise ChironError("Input data cannot be None")
        
        # Additional security validations would go here
        # - Input sanitization
        # - Schema validation
        # - Size limits
        # - Content filtering